+++ lib/internal/Magento/Framework/Mail/Template/TransportBuilder.php
@@ -279,4 +279,20 @@ class TransportBuilder
 
         return $this;
     }
+
+    /**
+     * Set mail from address by store.
+     *
+     * @param string|array $from
+     * @param string|int $store
+     *
+     * @return $this
+     */
+    public function setFromByStore($from, $store)
+    {
+        $result = $this->senderResolver->resolve($from, $store);
+        $this->message->setFrom($result['email'], $result['name']);
+
+        return $this;
+    }
 }
